pyproject.toml
[project]
name = "03-program2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types>=0.7.0",
    "anyio>=4.10.0",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.116.1",
    "idna>=3.10",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.7",
    "python-jose>=3.5.0",
    "sniffio>=1.3.1",
    "sqlalchemy>=2.0.43",
    "starlette>=0.47.2",
    "typing-extensions>=4.14.1",
    "typing-inspection>=0.4.1",
    "wheel>=0.45.1",
]
